stop manually moving the cursor around
it looks like vim is super buggy with how this interacts with undo (it
seems to just corrupt even previous entries in the undo tree, somehow?).
this makes undo work again, but doesn't fix `.`, because all of these
<Esc> commands start new actions (it's odd that actions and undo blocks
aren't the same here, honestly), but i think i might be able to work
around that in a similar way to repeat.vim.